requirements.txt
streamlit
scikit-learn
pandas
numpy
nltk
transformers
torch
newspaper3k
requests
beautifulsoup4
